feat(gdn): generalize Phase6 integration to A2 A3 A5
