+++ docs/source/developer_guide/Design_Documents/index.md
@@ -1,3 +1,6 @@
 # Design Documents
 
-This section provides an overview of the features implemented in vLLM Ascend. Developers can refer to this guide to understand how vLLM Ascend works.
+This section provides an overview of the features implemented in vLLM Ascend.
+Developers can refer to this guide to understand how vLLM Ascend works.
+
+- [Qwen3.5/Qwen3.6 cross-SoC FLA GDN integration](qwen_gdn_fla.md)
