+++ docs/superpowers/plans/2026-08-25-qwen35-qwen36-gdn-a5.md
@@ -1,5 +1,12 @@
 # Qwen3.5/Qwen3.6 GDN A5 Implementation Plan
 
+> Historical Stage 1 plan. The implemented adapter was generalized to A2/A3/A5
+> on 2026-08-29 and its preferred prefill path became the Phase 6 fused core.
+> The current authoritative design is
+> `docs/superpowers/specs/2026-08-29-qwen35-qwen36-gdn-fla-design.md`.
+> This plan is retained only as an execution history; its goals, file names,
+> operator count, and acceptance boundary must not guide new work.
+
 > **For agentic workers:** REQUIRED SUB-SKILL: Use superpowers:subagent-driven-development (recommended) or superpowers:executing-plans to implement this plan task-by-task. Steps use checkbox (`- [ ]`) syntax for tracking.
 
 **Goal:** Add an A5-only GDN inference adapter that prefers nine `fla_npu` forward operators, safely falls back to current vLLM-Ascend operators, and exposes precise selection/failure logs and smoke tests for Qwen3.5/Qwen3.6.
@@ -8,7 +15,7 @@
 
 **Tech Stack:** Python, PyTorch/torch_npu, vLLM custom-op integration, `fla_npu.ops.ascendc`, `fla_npu.ops.triton`, pytest.
 
-**Spec:** `docs/superpowers/specs/2026-08-25-qwen35-qwen36-gdn-a5-design.md`
+**Current spec:** `docs/superpowers/specs/2026-08-29-qwen35-qwen36-gdn-fla-design.md`
 
 ## Global Constraints
