+++ tests/e2e/pull_request/two_card/test_qwen3_6_27b_fia.py
@@ -20,7 +20,7 @@ import pytest
 from vllm.assets.image import ImageAsset
 
 from tests.e2e.conftest import VllmRunner, qwen_prompt, wait_until_npu_memory_free
-from vllm_ascend.device.device_config import is_950
+from vllm_ascend.device.device_config import is_fla_gdn_supported
 
 #MODEL = "Qwen/Qwen3.6-27B"
 MODEL = os.environ.get(
@@ -72,10 +72,10 @@ def test_qwen3_6_27b_multimodel_fia_eager():
 
 
 @patch.dict(os.environ, {"HCCL_BUFFSIZE": "1024"})
-@pytest.mark.skipif(not is_950(), reason="A5-only GDN model smoke")
+@pytest.mark.skipif(not is_fla_gdn_supported(), reason="requires A2/A3/A5 FLA GDN support")
 @wait_until_npu_memory_free()
-def test_qwen3_6_27b_gdn_a5_eager_smoke():
-    """Compare the strict A5 GDN adapter against the native eager path."""
+def test_qwen3_6_27b_gdn_fla_eager_smoke():
+    """Compare the strict FLA GDN adapter against the native eager path."""
     image = ImageAsset("cherry_blossom").pil_image.convert("RGB")
     prompts = qwen_prompt(["What is the content of this image?"])
     outputs_by_backend = {}
