+++ tests/ut/device/test_device_config.py
@@ -7,6 +7,8 @@ from vllm_ascend.device.device_config import (
     check_ascend_device_type,
     get_ascend_device_type,
     get_device_config,
+    get_fla_gdn_soc,
+    is_fla_gdn_supported,
 )
 from vllm_ascend.device.hardware import (
     AscendDeviceType,
@@ -52,6 +54,30 @@ def test_device_config_uses_build_info(monkeypatch):
     monkeypatch.setattr(_build_info, "__device_type__", "A3")
 
     assert get_ascend_device_type() is AscendDeviceType.A3
+
+
+@pytest.mark.parametrize(
+    ("device_type", "expected_soc"),
+    [
+        ("A2", "ascend910b"),
+        ("A3", "ascend910_93"),
+        ("A5", "ascend950"),
+    ],
+)
+def test_fla_gdn_is_supported_on_accelerator_families(monkeypatch, device_type, expected_soc):
+    """Catch regressions that restrict the shared FLA GDN operator to A5."""
+    monkeypatch.setattr(_build_info, "__device_type__", device_type)
+
+    assert is_fla_gdn_supported()
+    assert get_fla_gdn_soc() == expected_soc
+
+
+def test_fla_gdn_is_not_enabled_on_310p(monkeypatch):
+    """Keep unsupported hardware on the existing native GDN path."""
+    monkeypatch.setattr(_build_info, "__device_type__", "_310P")
+
+    assert not is_fla_gdn_supported()
+    assert get_fla_gdn_soc() is None
 
 
 def test_device_config_supports_legacy_soc_build_info(monkeypatch):
