+++ vllm_ascend/device/device_config.py
@@ -54,6 +54,25 @@ def get_ascend_device_type() -> AscendDeviceType:
     return get_device_config()._device_type
 
 
+_FLA_GDN_SOC_BY_DEVICE_TYPE = {
+    AscendDeviceType.A2: "ascend910b",
+    AscendDeviceType.A3: "ascend910_93",
+    AscendDeviceType.A5: "ascend950",
+}
+
+
+def get_fla_gdn_soc() -> str | None:
+    """Return the FLA build target for hardware that supports the GDN operators."""
+
+    return _FLA_GDN_SOC_BY_DEVICE_TYPE.get(get_ascend_device_type())
+
+
+def is_fla_gdn_supported() -> bool:
+    """Whether the current hardware family supports the shared FLA GDN path."""
+
+    return get_fla_gdn_soc() is not None
+
+
 def is_310p() -> bool:
     return get_ascend_device_type() == AscendDeviceType._310P
