+++ vllm_ascend/envs.py
@@ -101,8 +101,8 @@ env_variables: dict[str, Callable[[], Any]] = {
     # "1": force enable, "0": force disable, None: auto-detect from CANN headers.
     "VLLM_ASCEND_ENABLE_BATCH_MEMCPY": lambda: os.getenv("VLLM_ASCEND_ENABLE_BATCH_MEMCPY", None),
     # Select the Qwen GDN operator backend. Valid values are "auto",
-    # "fla_npu", and "native". A5 defaults to auto selection; non-A5
-    # devices keep using the native implementation.
+    # "fla_npu", and "native". A2, A3, and A5 default to auto selection;
+    # unsupported devices keep using the native implementation.
     "VLLM_ASCEND_GDN_BACKEND": lambda: os.getenv("VLLM_ASCEND_GDN_BACKEND", "auto").lower(),
     # Optional comma-separated per-operator GDN backend overrides, for example
     # "causal_conv1d=fla_npu,chunk_fwd_o=native". Values must be either
